extract Enum classes into new file, adjust imports and usages, and update preparations & categories to reutrn id values.
