+++ galley/enums.py
@@ -1,0 +1,16 @@
+from enum import Enum
+
+class MenuCategoryEnum(Enum):
+    """
+    Enum for categories, for item type menu <category name>: <category id>
+    """
+    MENU_TYPE = 'Y2F0ZWdvcnk6MjQ2NQ=='
+
+
+class PreparationEnum(Enum):
+    """
+    Enum for preparations.  <preparation name>: <preparation id>
+    """
+    STANDALONE = 'cHJlcGFyYXRpb246MjgzMzQ='
+    TWO_OUNCE_RAM = 'cHJlcGFyYXRpb246MjgxMTU='
+    THREE_OUNCE_RAM = 'cHJlcGFyYXRpb246MjgxMTQ='
